% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{boostrapCCA}
\alias{boostrapCCA}
\title{Bootstrap to get empirical intervals of canonical correlation.}
\usage{
boostrapCCA(
  A,
  B,
  nonzero_a,
  nonzero_b,
  cancor,
  folds = 10,
  n = 100,
  ci_quant = 0.01,
  silent = TRUE,
  toPlot = FALSE,
  parallel_logic = TRUE,
  nuisanceVar = 0,
  testStatType = "CC"
)
}
\arguments{
\item{A, B}{Data matrices.}

\item{nonzero_a, nonzero_b}{Numeric. Scalar or vector over the number of nonzeroes allowed for a correlation estimate.}

\item{cancor}{Numeric. Scalar or vector: anonical correlation estimate(s).}

\item{folds}{Numeric. Number of folds for the cross-validation process.}

\item{n}{Numeric. Times of bootstrapping.}

\item{ci_quant}{Numeric. Between 0 and 1. quantile of intervale.}

\item{silent}{Logical. If FALSE, a progress bar will appear on the console. Default is FALSE.}

\item{toPlot}{Logical. If TRUE, plot will be generated automatically showing the estimated canonical weights. Default is TRUE.}

\item{parallel_logic}{Logical. If TRUE, cross-validation is done in parallel.Default is FALSE.}

\item{nuisanceVar}{Data with nuisance variables. For statistic type.}

\item{testStatType}{Character. Choice of statistic. Options are CC (default), Wilks and Roy.}

\item{K}{Numeric. Number of components to be computed.}

\item{draws}{Numeric. Number of permutations for each component.}
}
\value{
Matrix with permutation estimates.
}
\description{
Bootstrap to get empirical intervals of canonical correlation.
}
\details{
For a exploratory analysis nonzero_a and nonzero_b can be vectors. The algorithm will then search for the best combination of sparsity choice nonzero_a and nonzero_b for each component.
}
