% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CCAtStat}
\alias{CCAtStat}
\title{Get the estatistic for the permutations.}
\usage{
CCAtStat(cancor, A, B, C = 0, type = c("CC", "Wilks", "Roy"))
}
\arguments{
\item{cancor}{Numeric. Canonical Correlation estimate.}

\item{A}{An nxp matrix.}

\item{B}{An nxq matrix.}

\item{C}{An nxs matrix. Confounding variables.}

\item{type}{Character. Choice of statistic: Canonical correlation, Wilks'statistic or Roy's statistic.}
}
\value{
Statistic
}
\description{
Get the estatistic for the permutations.
}
