% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cpev.fun}
\alias{cpev.fun}
\title{Calculated cummulative percentage of explained variance.}
\usage{
cpev.fun(mat, matK)
}
\arguments{
\item{mat}{An nxp matrix.}

\item{maxK}{An nxk matrix. Each column corresponds to a latent variable.}
}
\value{
Scalar.
}
\description{
Calculated cummulative percentage of explained variance.
}
