% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fastEigen}
\alias{fastEigen}
\title{Performs eigen decomposition of a matrix in PS space.}
\usage{
fastEigen(A)
}
\arguments{
\item{A}{A square matrix nxn.}
}
\value{
Matrix. Positive definite matrix.
}
\description{
Performs eigen decomposition of a matrix in PS space.
}
