% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{KFoldSCCA}
\alias{KFoldSCCA}
\title{Sparse Canonical Correlation Analysis. Computation of CC via NIPALS with soft thresholding.}
\usage{
KFoldSCCA(
  A,
  B,
  nonzero_a,
  nonzero_b,
  alpha_init,
  folds = 1,
  parallel_logic = FALSE,
  silent = FALSE,
  toPlot = TRUE,
  ATest_res = NULL,
  BTest_res = NULL
)
}
\arguments{
\item{A, B}{Data matrices.}

\item{nonzero_a, nonzero_b}{Numeric. Scalar or vector over the number of nonzeroes allowed for a correlation estimate.}

\item{silent}{Logical. If FALSE, a progress bar will appear on the console. Default is FALSE.}

\item{alphaInit}{Character. Type initialisation for \deqn{\mathbf{\alpha}}. Default is "eigen".}

\item{iter}{Numeric. Maximum number of iterations. Default is 20.}

\item{tol}{Numeric. Tolerance threshold. Default is 10^6.}
}
\value{
a list with the following elements:
\itemize{
\item{alpha}{Canonical vector for matrix \deqn{\mathbf{A}}, for each combination of sparsity value specified.}
\item{beta}{Canonical vector for matrix \deqn{\mathbf{B}}, for each combination of sparsity value specified.}
\item{cancor}{Max. canonical correlation estimate.}
\item{nonzero_a,nonzero_b}{Optimal nonzero values for each canonical vector.}
}
\description{
Sparse Canonical Correlation Analysis. Computation of CC via NIPALS with soft thresholding.
}
