% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{myHeatmap}
\alias{myHeatmap}
\title{Plot heatmap of cv w.r.t. the penalty parameter perfotmance.}
\usage{
myHeatmap(
  mat,
  palette_values = mpalette,
  blue = NULL,
  xlab = "",
  ylab = "",
  show_axes = TRUE,
  show_labels = TRUE,
  K = NULL
)
}
\arguments{
\item{mat}{A matrix.}

\item{palette_values}{Character string. Vector of colour values for the heatmap. Default is package's palette.}

\item{blue}{Logical. If TRUE, use only scale of blues from palette.}

\item{xlab}{Character. Label for x axis.}

\item{ylab}{Character. Label for y axis.}

\item{show_axes}{Logic. Default is True.}

\item{show_labels}{Logic. Default is True.}

\item{K}{Numeric. Number of components.}
}
\value{
No return value, called for plotting heatmap.
}
\description{
This function plots cca for different thresholds
}
\examples{
mat <- matrix(rexp(200, rate=.1), ncol=20)/200
myHeatmap(mat)
}
