% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{powerMethod}
\alias{powerMethod}
\title{Performs power method.}
\usage{
powerMethod(mat, vec, tol = 10^(-6), maxIter = 500, silent = TRUE)
}
\arguments{
\item{mat}{A square matrix nxn.}

\item{vec}{A vector of dimensions nx1.}

\item{tol}{Convergence criterion. Default is 10^(-6).}

\item{maxIter}{Maximum iterations. Default is 500.}

\item{silent}{Logical. If TRUE, convergence performance will be printed.}
}
\value{
List: vec: eigen vector; lambda: eigen value; t: total iterations.
}
\description{
Performs power method.
}
