% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{progressBar}
\alias{progressBar}
\title{Progress bar}
\usage{
progressBar(end, round)
}
\arguments{
\item{end}{maximum number of times a process will run.}

\item{round}{current round}
}
\value{
Display in consol of current status.
}
\description{
Shows progress of a process.
}
