% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scaledResidualMat}
\alias{scaledResidualMat}
\title{Performs scalling for matrix residualisation based on calculated coefficients.}
\usage{
scaledResidualMat(A)
}
\arguments{
\item{A}{An nxp matrix.}
}
\value{
scaled matrix.
}
\description{
Performs scalling for matrix residualisation based on calculated coefficients.
}
