% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{summary.toscca_object}
\alias{summary.toscca_object}
\title{Plot heatmap of cv w.r.t. the penalty parameter perfotmance.}
\usage{
\method{summary}{toscca_object}(object, ...)
}
\arguments{
\item{object}{toscca object.}

\item{...}{further arguments passed to or from methods.}
}
\value{
No return value, called for plotting heatmap.
}
\description{
This function plots cca for different thresholds
}
\examples{
\donttest{
# example code
#sample size etc
N = 10
p = 25
q = 5
# noise
X0 = sapply(1:p, function(x) rnorm(N))
Y0 = sapply(1:q, function(x) rnorm(N))

colnames(X0) = paste0("x", 1:p)
colnames(Y0) = paste0("y", 1:q)

# signal
Z1 = rnorm(N,0,10)


#Some associations with the true signal
alpha = (6:10) / 10
beta  = -(2:3) / 10

loc_alpha = 1:length(alpha)
loc_beta  = 1:length(beta)

for(j in 1:length(alpha))
  X0[, loc_alpha[j]] =  alpha[j] * Z1 + rnorm(N,0,0.03)

for(j in 1:length(beta))
  Y0[, loc_beta[j]] =  beta[j] * Z1 + rnorm(N,0,0.03)

# performa toscca
X = standardVar(X0)
Y = standardVar(Y0)
K = 2                                       # number of components to be estimated
nonz_x = c(2,5, 10, 20)                     # number of nonzero variables for X
nonz_y = c(2, 3, 4)                      # number of nonzero variables for Y
init   = "uniform"                          # type of initialisation
cca_toscca  = toscca(X, Y, nonz_x, nonz_y, K, alpha_init = init, type =1,
combination = TRUE, K=1, silent = TRUE, toPlot = FALSE)
summary(cca_toscca)
}
}
