% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{toscca.lv}
\alias{toscca.lv}
\title{Get latent variables}
\usage{
toscca.lv(data, alpha, beta)
}
\arguments{
\item{data}{List containint both observation amtrices.}

\item{alpha}{px1 numeric vector. canonical weights for X.}

\item{beta}{qx1 numeric vector. canonical weights for Y.}
}
\description{
Gets latent variables from data and estimates.
}
