% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tosccamm}
\alias{tosccamm}
\title{Computes TOSCCA-MM}
\usage{
tosccamm(
  A,
  B,
  nonzero_a,
  nonzero_b,
  folds = 1,
  parallel_logic = FALSE,
  toPlot = TRUE,
  silent = FALSE,
  ATest_res = NULL,
  BTest_res = NULL,
  model = "lme",
  lmeformula = " ~ -1 + time + (1|id)",
  predictor = NULL,
  arformula = c(1, 0, 0)
)
}
\arguments{
\item{A}{A data.frame with id and time as first two columns.}

\item{B}{A data.frame with id and time as first two columns.}

\item{nonzero_a}{Integer. Threshold parameter for A.}

\item{nonzero_b}{Integer. Threshold parameter for B.}

\item{folds}{Integer. Indicates number of folds to perform.}

\item{parallel_logic}{Logical. TRUE to parallelise folds. Default is FALSE.}

\item{toPlot}{Logical. If TRUE, plot will be generated automatically showing the estimated canonical weights. Default is TRUE.}

\item{silent}{Logical. TRUE to keep silent output messages. Default is FALSE.}

\item{ATest_res}{NULL. Keep NULL.}

\item{BTest_res}{NULL. Keep NULL.}

\item{model}{Character. c("lme", "ar"). Model to fit longitudinal latent space.}

\item{lmeformula}{Character. LME formula. Default is " ~ -1 + time + (1|id)".}

\item{predictor}{data.frame object with variables id and/or time to be used in lmeformula.}

\item{arformula}{Numeric vector. Choice of ARIMA. Default is c(1,0,0).}
}
\value{
List with estimated tosccamm parameters.
}
\description{
This function estimates sparse canonical vectors for matrices with multiple measurements
and the trajectories of the latent variables.
}
