% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{cpev.toscca}
\alias{cpev.toscca}
\title{Calculates cpev}
\usage{
cpev.toscca(mat, weights)
}
\arguments{
\item{mat}{A matrix.}

\item{weights}{A numberic vector of canonical weights.}
}
\value{
Returns cpev values for the kth or 1:K components.
}
\description{
This function calculates cpev
w.r.t. a chosen origin.
}
