% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{getWhich}
\alias{getWhich}
\title{Get location of required.}
\usage{
getWhich(data, fun)
}
\arguments{
\item{data}{Numeric matrix.}

\item{fun}{Function to search data.}
}
\value{
Returns value matching function fun.
}
\description{
Get location of required.
}
\examples{
getWhich(rnorm(100), max)

}
