% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{initialiseCanVar}
\alias{initialiseCanVar}
\title{Initialised the canonical vector for the iterative process based on
positive eigen values. Then, SVD is performed on that PS matrix.}
\usage{
initialiseCanVar(A, B)
}
\arguments{
\item{A}{An nxp matrix.}

\item{B}{An nxq matrix.}
}
\value{
An pzp vector.
}
\description{
Initialised the canonical vector for the iterative process based on
positive eigen values. Then, SVD is performed on that PS matrix.
}
