% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{myHeatmap}
\alias{myHeatmap}
\title{Plot heatmap}
\usage{
myHeatmap(mat, palette = "Teal", coln = 12, xlab = "", ylab = "", axes = FALSE)

myHeatmap(mat, palette = "Teal", coln = 12, xlab = "", ylab = "", axes = FALSE)
}
\arguments{
\item{mat}{A square matrix nxm.}

\item{palette}{Choice of palette. Default is Teal.}

\item{coln}{Integer. Grid length parameter.}

\item{xlab}{Character. Label for x axis.}

\item{ylab}{Character. Label for y axis.}

\item{axes}{Logical. If TRUE, axes are between 0 and 1.}

\item{pallette}{Character. Name of pallette colour. Default is Teal.}
}
\value{
Grid plot.

Grid plot.
}
\description{
This function generated a heatmap. Withing the package it is used to
provide with a visualisation of the exploration of optimal sparsity levels.
}
