% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{myHeatmap}
\alias{myHeatmap}
\title{Plots grid of CC over threshold penalty}
\usage{
myHeatmap(mat, palette = "Teal", coln = 12, xlab = "", ylab = "", axes = FALSE)
}
\arguments{
\item{mat}{A square matrix nxm.}

\item{coln}{Integer. Grid length parameter.}

\item{xlab}{Character. Label for x axis.}

\item{ylab}{Character. Label for y axis.}

\item{axes}{Logical. If TRUE, axes are between 0 and 1.}

\item{pallette}{Character. Name of pallette colour. Default is Teal.}
}
\value{
Grid plot.
}
\description{
Plots grid of CC over threshold penalty
}
