% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{myHeatmap}
\alias{myHeatmap}
\title{Plot heatmap}
\usage{
myHeatmap(mat, palette = "Teal", coln = 12, xlab = "", ylab = "", axes = FALSE)
}
\arguments{
\item{mat}{Matrix containing values along a grid.}

\item{palette}{Choice of palette. Default is Teal.}

\item{coln}{Number of columns for the grid distribution. Default is 12.}

\item{xlab}{Lable for X axis.}

\item{ylab}{Lable for Y axis.}

\item{axes}{Logical. Have axes between 0 and 1. Default is FALSE.}
}
\value{
Grid plot.
}
\description{
This function generated a heatmap. Withing the package it is used to
provide with a visualisation of the exploration of optimal sparsity levels.
}
