% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{plt.selstab}
\alias{plt.selstab}
\title{Plot slection stability for penalty parameter performance.}
\usage{
plt.selstab(object, X, Y, palette_values = mpalette, blue = T, mm = T, k = 1)
}
\arguments{
\item{object}{A toscca object.}

\item{X}{nxp matrix. Observation matrix.}

\item{Y}{A nxq matrix. Observation matrix.}

\item{palette_values}{Character. Name of a palette for the heatmap. Default is "Teal".}

\item{blue}{Logical. If TRUE, use only scale of blues from palette.}

\item{mm}{Logic. Indicates whether there are multiple measurements or not. Default is True.}

\item{k}{Numeric. Component, default k =1.}
}
\description{
This function plots cv for different thresholds
}
\examples{
# example code
\dontrun{
plt.selstab(mod,X, Y)
}
}
