% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{plt.selstab}
\alias{plt.selstab}
\title{Plot slection stability for penalty parameter performance.}
\usage{
plt.selstab(
  object,
  X,
  Y,
  palette_values = mpalette,
  blue = TRUE,
  mm = TRUE,
  k = 1
)
}
\arguments{
\item{object}{A toscca object.}

\item{X}{nxp matrix. Observation matrix.}

\item{Y}{A nxq matrix. Observation matrix.}

\item{palette_values}{Character. Name of a palette for the heatmap. Default is "Teal".}

\item{blue}{Logical. If TRUE, use only scale of blues from palette.}

\item{mm}{Logic. Indicates whether there are multiple measurements or not. Default is True.}

\item{k}{Numeric. Component, default k =1.}
}
\value{
No return value, called for selection stability pot.
}
\description{
This function plots cv for different thresholds
}
\examples{
# example code
\dontrun{
plt.selstab(object,X, Y)
}
}
