% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{standardVar}
\alias{standardVar}
\title{Stardardise a matrix}
\usage{
standardVar(mat, centre = TRUE, normalise = FALSE)
}
\arguments{
\item{centre}{Logical, if true, cetre to mean zero.}

\item{normalise}{Logical, if true, performs vector normalisation.}

\item{X}{Matrix or vector to be standardise.}
}
\value{
A matrix or vector with the preferred standardarisation
}
\description{
This function stardardises a matrix or a vector and gives the option to
centre or normalise (only vectors).
}
